Preserve old comment behaviour in CommentBiff7
